All the prefs!
This commit:

- Closes #34 by reading pref values from their own config object (not a
JSON file, though).
- Closes #54 by setting prefs for hiding the TP and FB UIs when they
should not be exposed to users.
- Closes #76 by setting security.pki.distrust_ca_policy to 1 everywhere.
- Closes #77 by opting out of the study if the user has a modified pref
value (and also cleaning up prefs properly).

The only thing I'm slightly unsure about is #32, so we should discuss
this and potentially move to a follow-up. I'm generally a bit uncertain
about how the weighting works, but luckily that should be easy to
configure.