The big content/frame script refactor
This commit removes the content script and instead uses the frame script
for all page communication.

Even with the added bureaucracy of going through api.js, this simplifies
things and hopefully stabilizes the way we send this data to the main
process:

- This makes us report vital information such as eTLD and page_reloaded as
soon as possible and guarantees that we get data even if the "load"
event was not fired.

- This should hopefully remove the uncertainties about whether our
content script had enough time to send the final unload message before
getting destroyed, since frame scripts stick around for the tab
lifetime.

- This makes us 100% certain that we always report the correct number of
trackers, since we now query them on beforeunload. I was previously
worried about race condition between onSecurityChange events and when
numTrackersFound is set.

- This enables us to easily use the eTLD service for getting the actual
eTLD now.