Make the reload survey apply to the document that was actually reloaded.
This introduces a new attribute on the tabInfo called
"payloadWaitingForSurvey", which keeps a payload that is currently held
back from sending because the user still needs to respond to the reload
survey.

During testing I have also found that the onBeforeUnload event tends to
be unreliable in certain cases (mostly when closing tabs), so this patch
further makes onUnload record data as well.. It was necessary to send some
of it along anyway, because the Browser:Reload event seems to arrive after
onbeforeunload and so we need to send it in onunload. Sorry for the
mess.