Get the document from the unload event instead of the docShell
From my testing, getting the document from the event target of the
unload event is more reliable than accessing the document getter on the
docShell. I'm not sure why the docShell loses this property earlier, but
it makes sense to use the event target since this is a web API and
should have more stability guarantees than our internal docShell.